feat: Rendering pieces - Add algorith to render pieces in the board - Change some data types to adequate with Fyne pattern

+++ figures/circle.go
@@ -8,7 +8,7 @@ import (
 )
 
 type Circle struct {
-	size  int
+	size  float32
 	color color.Color
 }
 
@@ -16,7 +16,7 @@ func NewCircle() Figure {
 	return &Circle{}
 }
 
-func (c *Circle) SetSize(size int) {
+func (c *Circle) SetSize(size float32) {
 	c.size = size
 }
 
@@ -27,6 +27,6 @@ func (c *Circle) SetColor(color color.Color) {
 func (c *Circle) Draw() fyne.CanvasObject {
 	circle := canvas.NewCircle(c.color)
 
-	circle.Resize(fyne.NewSize(float32(c.size), float32(c.size)))
+	circle.Resize(fyne.NewSize(c.size, c.size))
 	return circle
 }
